refactored some of the styling by creating a shared styling folder at root
